UI: move the codename badge renderer to badge.rs
Structure-only commit: codename_badge_lines and ops_color join the
braille assets they render, so badge.rs owns the whole badge and
page.rs is purely page composition and column layout.
